Create README.md
This program inputs a set of character from a text file
Keys.txt and then processes the the text line by line
and presses certain keys on the 'virtual' keyboard with
a delay of 400 milli-seconds and then it ends the program
There are certain problems with the ending due to the
structure of the loops but the program is by all means
very reliable as long as the text file is not messed up
by the user. The keyboard combinations which result in
the pressing of a corresponding key are given in the source
code feel free to share.